refactor: 게임 횟수 로직 수정 및 랭킹 산정 시 redisRepository 초기화 및 저장하도록 수정

+++ core/core-infra-redis/src/main/java/com/pgms/coreinfraredis/repository/RedisInGameRepository.java
@@ -84,9 +84,9 @@ public class RedisInGameRepository {
 	}
 
 	// 누적 멤버 점수 조회
-	public Long getTotalScore(String roomId, String memberId) {
+	public Double getTotalScore(String roomId, String memberId) {
 		Double score = redisTemplate.opsForZSet().score(TOTAL_PREFIX + roomId, memberId);
-		return score != null ? score.longValue() : 0;
+		return score != null ? score.longValue() : 0.0;
 	}
 
 	public Map<Long, Double> getTotalScores(String roomId) {
